requirements.txt
asgiref==3.3.1
da-vinci==0.3.0
dj-database-url==0.5.0
Django==3.1.4
django-ajax-datatable==4.1.4
django-appconf==1.0.4
django-avatar==5.0.0
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.10.0
django-filter==2.4.0
django-mathfilters==1.0.0
django-messages==0.6.0
django-template-maths==0.2.0
django-thumbnails==0.4.0
django-thumbs==0.4
django-thumbs-v2==0.4.1
djangorestframework==3.12.2
gunicorn==20.0.4
mysqlclient==2.0.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
reportlab==3.5.56
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
